will ADD and REFACTOR: minor styling changes and refactoring the entire codebase, sorting files into appropriate folders
