+++ solar-panel-planner/package.json
@@ -13,7 +13,6 @@
     "@react-google-maps/api": "^2.20.3",
     "@reduxjs/toolkit": "^2.3.0",
     "@vis.gl/react-google-maps": "^1.4.0",
-    "axios": "^1.7.9",
     "esbuild": "^0.24.0",
     "framer-motion": "^11.12.0",
     "jspdf": "^2.5.2",
