+++ solar-panel-planner/src/ui/Modal.jsx
