+++ solar-panel-planner/src/ui/ShowAvailableTimeSlot.jsx
