+++ solar-panel-planner/src/ui/ShowMap.jsx
