+++ solar-panel-planner/src/utils/VisitExport.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import { useState } from "react";
 import { useSelector } from "react-redux";
 // import { useJsApiLoader } from "@react-google-maps/api";
 import moment from "moment";
@@ -93,7 +93,7 @@ export default function VisitList() {
       <button
         onClick={getOptimizedRoute}
         className="bg-primaryGreen text-white px-4 py-2 rounded hover:bg-secondaryGreen">
-        Export today's route
+        Export today&apos;s route
       </button>
     </div>
   )
