+++ solar-panel-planner/src/utils/userDb.js
